issue #21 Slightly refactor face analyser
