+++ sinner/FaceAnalyser.py
@@ -6,19 +6,22 @@ from sinner.typing import Frame, Face
 
 
 class FaceAnalyser:
-    THREAD_LOCK = threading.Lock()
-
-    _face_analyser: FaceAnalysis
+    _face_analyser: FaceAnalysis | None = None
     _execution_providers: List[str]
 
     def __init__(self, execution_providers: List[str]):
         self._execution_providers = execution_providers
-        with self.THREAD_LOCK:
-            self._face_analyser = FaceAnalysis(name='buffalo_l', providers=self._execution_providers)
-            self._face_analyser.prepare(ctx_id=0, det_size=(640, 640))
+
+    @property
+    def face_analyser(self) -> FaceAnalysis:
+        if self._face_analyser is None:
+            with threading.Lock():
+                self._face_analyser = FaceAnalysis(name='buffalo_l', providers=self._execution_providers)
+                self._face_analyser.prepare(ctx_id=0, det_size=(640, 640))
+        return self._face_analyser
 
     def get_one_face(self, frame: Frame) -> None | Face:
-        face = self._face_analyser.get(frame)
+        face = self.face_analyser.get(frame)
         try:
             return min(face, key=lambda x: x.bbox[0])
         except ValueError:
@@ -26,6 +29,6 @@ class FaceAnalyser:
 
     def get_many_faces(self, frame: Frame) -> None | List[Face]:
         try:
-            return self._face_analyser.get(frame)
+            return self.face_analyser.get(frame)
         except IndexError:
             return None
